+++ magic-compiler/src/stdlib/clojure/core_deftype.clj
@@ -544,7 +544,7 @@
 (defn- expand-method-impl-cache [^clojure.lang.MethodImplCache cache c f]
   (if (.map cache)
     (let [cs (assoc (.map cache) c (clojure.lang.MethodImplCache+Entry. c f))]                            ;;; clojure.lang.MethodImplCache$Entry
-      (clojure.lang.MethodImplCache. (.protocol cache) (.methodk cache) cs))
+      (clojure.lang.MethodImplCache. (.sym cache) (.protocol cache) (.methodk cache) cs))
     (let [cs (into1 {} (remove (fn [[c e]] (nil? e)) (map vec (partition 2 (.table cache)))))
           cs (assoc cs c (clojure.lang.MethodImplCache+Entry. c f))]                                      ;;; clojure.lang.MethodImplCache$Entry
       (if-let [[shift mask] (maybe-min-hash (map hash (keys cs)))]
@@ -555,8 +555,8 @@
                                  (aset t (inc i) e)
                                  t))
                              table cs)]
-          (clojure.lang.MethodImplCache. (.protocol cache) (.methodk cache) shift mask table))
-        (clojure.lang.MethodImplCache. (.protocol cache) (.methodk cache) cs)))))
+          (clojure.lang.MethodImplCache. (.sym cache) (.protocol cache) (.methodk cache) shift mask table))
+        (clojure.lang.MethodImplCache. (.sym cache) (.protocol cache) (.methodk cache) cs)))))
 
 (defn- super-chain [^Type c]           ;;; Class
   (when c
@@ -620,7 +620,7 @@
     (set! (.__methodImplCache pf) (expand-method-impl-cache cache (class x) f))
     f))
 
-(defn- emit-method-builder [on-interface method on-method arglists]
+(defn- emit-method-builder [on-interface method on-method arglists extend-via-meta]
   (let [methodk (keyword method)
         gthis (gensym)
         gthis-tagged (with-meta gthis {:tag 'clojure.lang.AFunction})
@@ -628,7 +628,7 @@
     `(fn [cache#]
       (let [~ginterf
              (fn
-               ~@(map 
+               ~@(map
                   (fn [args]
                     (let [gargs (-> (map #(gensym (str "gf__" % "__")) args)
                                     vec
@@ -639,23 +639,34 @@
                   arglists))
              ^clojure.lang.AFunction f#
              (fn ~gthis
-               ~@(map 
+               ~@(map
                   (fn [args]
                     (let [gargs (map #(gensym (str "gf__" % "__")) args)
                           target (first gargs)]
-                      `([~@gargs]
-                          (let [cache# (.__methodImplCache ~gthis-tagged)
-                                f# (.fnFor cache# (clojure.lang.Util/classOf ~target))]
-                            (if f# 
-                              (f# ~@gargs)
-                              ((-cache-protocol-fn ~gthis ~target ~on-interface ~ginterf) ~@gargs))))))
+                      (if extend-via-meta
+                        `([~@gargs]
+                            (let [cache# (.__methodImplCache ~gthis-tagged)
+                                  f# (.fnFor cache# (clojure.lang.Util/classOf ~target))]
+                              (if (identical? f# ~ginterf)
+                                (f# ~@gargs)
+                                (if-let [meta# (when-let [m# (meta ~target)] ((.sym cache#) m#))]
+                                  (meta# ~@gargs)
+                                  (if f#
+                                    (f# ~@gargs)
+                                    ((-cache-protocol-fn ~gthis ~target ~on-interface ~ginterf) ~@gargs))))))
+                        `([~@gargs]
+                            (let [cache# (.__methodImplCache ~gthis-tagged)
+                                  f# (.fnFor cache# (clojure.lang.Util/classOf ~target))]
+                              (if f#
+                                (f# ~@gargs)
+                                ((-cache-protocol-fn ~gthis ~target ~on-interface ~ginterf) ~@gargs)))))))
                   arglists))]
          (set! (.__methodImplCache f#) cache#)
          f#))))
 
 (defn -reset-methods [protocol]
   (doseq [[^clojure.lang.Var v build] (:method-builders protocol)]
-    (let [cache (clojure.lang.MethodImplCache. protocol (keyword (.sym v)))]
+    (let [cache (clojure.lang.MethodImplCache. (symbol v) protocol (keyword (.sym v)))]
       (.bindRoot v (build cache)))))
 
 (defn- assert-same-protocol [protocol-var method-syms]
@@ -725,7 +736,8 @@
                                 (mapcat 
                                  (fn [s]
                                    [`(intern *ns* (with-meta '~(:name s) (merge '~s {:protocol (var ~name)})))
-                                    (emit-method-builder (:on-interface opts) (:name s) (:on s) (:arglists s))])
+                                    (emit-method-builder (:on-interface opts) (:name s) (:on s) (:arglists s)
+                                                         (:extend-via-metadata opts))])
                                  (vals sigs)))))
      (-reset-methods ~name)
      '~name)))
@@ -737,6 +749,9 @@
     ;optional doc string
     \"A doc string for AProtocol abstraction\"
 
+   ;options
+   :extend-via-metadata true
+
   ;method signatures
     (bar [this a b] \"bar docs\")
     (baz [this a] [this a b] [this a b c] \"baz docs\"))
@@ -746,10 +761,17 @@
   polymorphic functions and a protocol object. All are
   namespace-qualified by the ns enclosing the definition The resulting
   functions dispatch on the type of their first argument, which is
-  required and corresponds to the implicit target object ('this' in 
-  Java parlance). defprotocol is dynamic, has no special compile-time 
-  effect, and defines no new types or classes. Implementations of 
+  required and corresponds to the implicit target object ('this' in
+  Java parlance). defprotocol is dynamic, has no special compile-time
+  effect, and defines no new types or classes. Implementations of
   the protocol methods can be provided using extend.
+
+  When :extend-via-metadata is true, values can extend protocols by
+  adding metadata where keys are fully-qualified protocol function
+  symbols and values are function implementations. Protocol
+  implementations are checked first for direct definitions (defrecord,
+  deftype, reify), then metadata definitions, then external
+  extensions (extend, extend-type, extend-protocol)
 
   defprotocol will automatically generate a corresponding interface,
   with the same name as the protocol, i.e. given a protocol: