+++ magic-compiler/test.clj
@@ -14,7 +14,9 @@
    magic.test.fn
    magic.test.letfn
    magic.test.pipeline
-   magic.test.hash)
+   magic.test.hash
+   magic.test.stdlib
+   magic.test.flags)
   (:use clojure.test))
 
 (defn all []
@@ -33,7 +35,9 @@
    'magic.test.fn
    'magic.test.letfn
    'magic.test.pipeline
-   'magic.test.hash))
+   'magic.test.hash
+   'magic.test.stdlib
+   'magic.test.flags))
 
 (defn run [& namespaces]
   (apply run-tests namespaces))